Add experimental AsyncImageSliceData
